+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Bugfix
 
 * fixes a bug where the pipeline index increases on every restart of a failed pipeline
+* fixes closed log queue issue by run logging in an extra process
 
 ## 11.0.0
 ### Breaking